+++ ytree/arbor/frontends/treefarm/__init__.py
@@ -1,5 +1,5 @@
 """
-tree_farm frontend
+treefarm frontend
 
 
 
@@ -13,5 +13,5 @@ tree_farm frontend
 # The full license is in the file COPYING.txt, distributed with this software.
 #-----------------------------------------------------------------------------
 
-from ytree.arbor.frontends.tree_farm.arbor import \
+from ytree.arbor.frontends.treefarm.arbor import \
     TreeFarmArbor
